requirements.txt
streamlit
SpeechRecognition
langchain-groq
gtts
streamlit_chat
Python_dotenv
